% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{check_build_publish}
\alias{check_build_publish}
\title{check - build and publish}
\usage{
check_build_publish(pkg = ".", binary = FALSE, path = tempdir(),
  repodir = getOption("dratRepo", "~/git/drat"), action = "none",
  commit = FALSE, vignettes = TRUE, manual = FALSE, args = NULL,
  quiet = FALSE, ...)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{binary}{Produce a binary (\code{--binary}) or source (
\code{--no-manual --no-resave-data}) version of the package.}

\item{path}{path in which to produce package.  If \code{NULL}, defaults to
the parent directory of the package.}

\item{repodir}{A local directory corresponding to the repository
top-level directory.}

\item{action}{A character string containing one of: \dQuote{none} 
(the default; add the new package into the repo, effectively masking 
previous versions), \dQuote{archive} (place any previous versions into 
a package-specific archive folder, creating such an archive if it does 
not already exist), or \dQuote{prune} (calling \code{\link{pruneRepo}}).}

\item{commit}{Either boolean toggle to select automatic git operations
\sQuote{add}, \sQuote{commit}, and \sQuote{push} or, alternatively,
a character variable can be used to specify a commit message; this also
implies the \sQuote{TRUE} values in other contexts.}

\item{vignettes}{For source packages: if \code{FALSE}, don't build PDF
vignettes (\code{--no-build-vignettes}) or manual (\code{--no-manual}).}

\item{manual}{For source packages: if FALSE, don't build PDF vignettes
(--no-build-vignettes) or manual (--no-manual).}

\item{args}{An optional character vector of additional command
line arguments to be passed to \code{R CMD build} if \code{binary = FALSE},
or \code{R CMD install} if \code{binary = TRUE}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{...}{furhter arguments passed through to devtools::check()}
}
\description{
check - build and publish
}
\seealso{
\link[devtools]{check} \link[devtools]{build} \link[drat]{insertPackage}
}

