% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{dratful}
\alias{dratful}
\title{the ultimative shortcut all-default function to check_build_publish()}
\usage{
dratful(pkg = ".", repodir = getOption("dratRepo", "~/git/drat"))
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{repodir}{A local directory corresponding to the repository
top-level directory.}
}
\description{
the ultimative shortcut all-default function to check_build_publish()
}

