% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dsc_content}
\alias{dsc_content}
\title{extrac package names and version from PACKAGES file}
\usage{
dsc_content(p_description)
}
\arguments{
\item{p_description}{content of a PACKAGES file}
}
\description{
extrac package names and version from PACKAGES file
}
\keyword{internal}

