% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{git_commit}
\alias{git_commit}
\title{git_commit}
\usage{
git_commit(repo, message = "-")
}
\arguments{
\item{repo}{a git2r::repository() object}

\item{message}{git commit message}
}
\description{
git_commit
}
\keyword{internal}

