% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{install_drat}
\alias{install_drat}
\title{installing from a github hosted drat and CRAN}
\usage{
install_drat(pkg, username = "ghrr")
}
\arguments{
\item{pkg}{name of the package to be installed}

\item{username}{github username}
}
\description{
installing from a github hosted drat and CRAN
}

