% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{make_index_files}
\alias{make_index_files}
\title{make index.html and README.md from PACKAGES files}
\usage{
make_index_files(repodir)
}
\arguments{
\item{repodir}{A local directory corresponding to the repository
top-level directory.}
}
\description{
make index.html and README.md from PACKAGES files
}
\keyword{internal}

