% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_dcf}
\alias{read_dcf}
\title{parsing DESCRIPTION files to lists}
\usage{
read_dcf(path = ".")
}
\arguments{
\item{path}{path to package DESCRIPTION}
}
\description{
parsing DESCRIPTION files to lists
}
\keyword{internal}

